+++ struct.h
@@ -9,16 +9,16 @@ typedef struct {
 //edge list structure:
 typedef struct {
 	unsigned long n;//number of nodes
-	unsigned long e;//number of edges
-	unsigned long emax;//max number of edges
+	unsigned long long e;//number of edges
+	unsigned long long emax;//max number of edges
 	edge *edges;//list of edges
-	unsigned long *cd;//cumulative degree cd[0]=0 length=n+1
+	unsigned long long *cd;//cumulative degree cd[0]=0 length=n+1
 	unsigned long *adj;//concatenated lists of neighbors of all nodes
 	unsigned long *map;//map[u]=original label of node u
 } adjlist;
 
 typedef struct {
-	unsigned n;//number of clusters
+	unsigned long n;//number of clusters
 	adjlist** sg;//sg[i]=pointer to adjlist of cluster i
 } clusters;
